+++ package.json
@@ -39,6 +39,7 @@
     "core-js": "^2.4.1",
     "dotenv": "^4.0.0",
     "history": "^4.6.3",
+    "hoist-non-react-statics": "^2.2.2",
     "i18next": "^8.4.3",
     "i18next-xhr-backend": "^1.4.2",
     "lodash": "^4.17.4",
